fix: refactoring code in repo

+++ app/Http/Controllers/ConfigurationController.php
@@ -2,18 +2,16 @@
 
 namespace App\Http\Controllers;
 
-use Auth;
+use App\Jobs\ProcessConfiguration;
 use App\Models\Configuration;
 use App\Models\User;
-use Illuminate\Http\Request;
-use Symfony\Component\Process\Process;
-use Symfony\Component\Process\Exception\ProcessFailedException;
-use Illuminate\Support\Facades\Storage;
+use Auth;
 use File;
-use App\Jobs\ProcessConfiguration;
-use Redirect;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
-
+use Redirect;
+use Symfony\Component\Process\Process;
 
 /**
  * Class ConfigurationController
@@ -36,13 +34,12 @@ class ConfigurationController extends Controller
     {
 
         $user = Auth::user();
-        $configurations =  $user->configurations();
+        $configurations = $user->configurations();
 
         return view('configuration.index', [
             'configurations' => $configurations->paginate(10),
         ]);
-
-       }
+    }
 
     /**
      * Show the form for creating a new resource.
@@ -74,9 +71,13 @@ class ConfigurationController extends Controller
 
             $port_detector = new Process(['python3.9', app_path('Runner/detector.py')]);
             $port_detector->run();
-            $assigned_port = $port_detector->getOutput();
 
-        } catch(Exception $e) {
+            if (!$port_detector->isSuccessful()) {
+                return redirect('configurations/create')->with('flama', 'We could not find a free port for your configuration, please try again later.')->withInput();
+            }
+
+            $assigned_port = $port_detector->getOutput();
+        } catch (Exception $e) {
 
             return redirect('configurations/create')->with('flama', 'No ports are available, try again later.')->withInput();
         }
@@ -103,10 +104,9 @@ class ConfigurationController extends Controller
                 'facebook_socials' => request('facebook_socials'),
                 'youtube_socials' => request('youtube_socials'),
                 'assigned_port' => $assigned_port,
-                'user_id' => Auth::user()->id
+                'user_id' => Auth::user()->id,
             ]);
-
-        } catch(Exception $e) {
+        } catch (Exception $e) {
 
             $configuration->delete();
             return redirect('configurations/create')->with('flama', 'Could not save the configuration, try again later.')->withInput();
@@ -133,44 +133,44 @@ class ConfigurationController extends Controller
         # Flama validation
         try {
 
-            File::makeDirectory("".$HOME_PATH."/webspl/app/Runner/websites/".$web_name."");
+            File::makeDirectory("" . $HOME_PATH . "/webspl/app/Runner/websites/" . $web_name . "");
 
-            $flama = new Process(['python3.9', app_path('Runner/flama.py'), $web_name, $admin_email, $theme, $php, $storage, $catalog, $search, $paypal_payment, $creditcard_payment,$mobile_payment, $cart, $security, $backup, $seo, $twitter_socials, $facebook_socials, $youtube_socials]);
+            $flama = new Process(['python3.9', app_path('Runner/flama.py'), $web_name, $admin_email, $theme, $php, $storage, $catalog, $search, $paypal_payment, $creditcard_payment, $mobile_payment, $cart, $security, $backup, $seo, $twitter_socials, $facebook_socials, $youtube_socials]);
             $flama->run();
+
+            if (!$flama->isSuccessful()) {
+                return redirect('configurations/create')->with('flama', 'Flama could not validate your configuration, please try again later.')->withInput();
+            }
 
             # Site builder
             try {
 
-                $line = fgets(fopen( "".$HOME_PATH."/webspl/app/Runner/websites/".$web_name."/result.txt", 'r'));
+                $line = fgets(fopen("" . $HOME_PATH . "/webspl/app/Runner/websites/" . $web_name . "/result.txt", 'r'));
                 if ($line == '1') {
 
                     ProcessConfiguration::dispatch($configuration, $username, $password);
-                
                 } else {
-        
-                    File::deleteDirectory("".$HOME_PATH."/webspl/app/Runner/websites/".$web_name."");
-                    $configuration->delete();
-                    return redirect('configurations/create')->with('flama', 'The generated configuration is not valid. Please, check the following constraints:')->withInput();           
-                }
 
-            } catch(Exception $e) {
-            
-                File::deleteDirectory("".$HOME_PATH."/webspl/app/Runner/websites/".$web_name."");
-                File::delete("".$HOME_PATH."/webspl/storage/".$username."/".$web_name.".zip");
+                    File::deleteDirectory("" . $HOME_PATH . "/webspl/app/Runner/websites/" . $web_name . "");
+                    $configuration->delete();
+                    return redirect('configurations/create')->with('flama', 'The generated configuration is not valid. Please, check the following constraints:')->withInput();
+                }
+            } catch (Exception $e) {
+
+                File::deleteDirectory("" . $HOME_PATH . "/webspl/app/Runner/websites/" . $web_name . "");
+                File::delete("" . $HOME_PATH . "/webspl/storage/" . $username . "/" . $web_name . ".zip");
                 $configuration->delete();
-    
+
                 return redirect('configurations/create')->with('flama', 'The configuration could not be generated, please try again later.')->withInput();
             }
+        } catch (Exception $e) {
 
-        } catch(Exception $e) {
-
-            File::deleteDirectory("".$HOME_PATH."/webspl/app/Runner/websites/".$web_name."");
+            File::deleteDirectory("" . $HOME_PATH . "/webspl/app/Runner/websites/" . $web_name . "");
             $configuration->delete();
             return redirect('configurations/create')->with('flama', 'The configuration could not be validated, please try again later.')->withInput();
         }
 
         return redirect()->route('configurations.index')->with('success', 'Configuration created successfully');
-        
     }
 
     /**
@@ -192,7 +192,7 @@ class ConfigurationController extends Controller
             $stopper = new Process(['python3.9', app_path('Runner/stopper.py'), $web_name]);
             $stopper->run();
 
-            if(!$stopper->isSuccessful()) {
+            if (!$stopper->isSuccessful()) {
 
                 return redirect()->route('configurations.index')->with('error', 'Your configuration could not be stopped, try again later.');
             }
@@ -200,8 +200,7 @@ class ConfigurationController extends Controller
             $configuration = Configuration::find($id);
             $configuration->status = 'PAUSED';
             $configuration->save();
-
-        } catch(Exception $e) {
+        } catch (Exception $e) {
 
             return redirect()->route('configurations.index')->with('error', 'Your configuration could not be stopped, try again later.');
         }
@@ -229,7 +228,7 @@ class ConfigurationController extends Controller
             $starter = new Process(['python3.9', app_path('Runner/starter.py'), $web_name]);
             $starter->run();
 
-            if(!$starter->isSuccessful()) {
+            if (!$starter->isSuccessful()) {
 
                 return redirect()->route('configurations.index')->with('error', 'Your configuration could not be started, please try again later.');
             }
@@ -237,8 +236,7 @@ class ConfigurationController extends Controller
             $configuration = Configuration::find($id);
             $configuration->status = 'READY';
             $configuration->save();
-
-        } catch(Exception $e) {
+        } catch (Exception $e) {
 
             return redirect()->route('configurations.index')->with('error', 'Your configuration could not be started, please try again later.');
         }
@@ -263,17 +261,16 @@ class ConfigurationController extends Controller
             $destroyer = new Process(['python3.9', app_path('Runner/destroyer.py'), $web_name]);
             $destroyer->run();
 
-            if(!$destroyer->isSuccessful()) {
+            if (!$destroyer->isSuccessful()) {
 
                 return redirect()->route('configurations.index')->with('error', 'Your configuration could not be deleted, try again later.');
             }
 
-            File::deleteDirectory("".$HOME_PATH."/webspl/app/Runner/websites/".$web_name."");
-            File::delete("".$HOME_PATH."/webspl/storage/app/".$username."/".$web_name.".zip");
+            File::deleteDirectory("" . $HOME_PATH . "/webspl/app/Runner/websites/" . $web_name . "");
+            File::delete("" . $HOME_PATH . "/webspl/storage/app/" . $username . "/" . $web_name . ".zip");
 
             $configuration = Configuration::find($id)->delete();
-
-        } catch(Exception $e) {
+        } catch (Exception $e) {
 
             return redirect()->route('configurations.index')->with('error', 'Your configuration could not be deleted, try again later.');
         }
